requirements.txt
aniso8601==3.0.2
atomicwrites==1.1.5
attrs==18.1.0
click==6.7
Flask==1.0.2
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.2.0
peewee==3.5.0
pluggy==0.6.0
py==1.5.3
python-slugify==1.2.5
pytz==2018.4
six==1.11.0
Unidecode==1.0.22
Werkzeug==0.14.1
